Removed Integer cast, as rates.get() returns an Integer, and unnecessary casting expressions make the code harder to read and understand

+++ WyCash/src/main/java/wycash/Bank.java
@@ -17,7 +17,7 @@ public class Bank {
 	public int rate(String from, String to) {
 		if (from.equals(to))
 			return 1;
-		Integer rate = (Integer) rates.get(new Pair(from, to));
+		Integer rate = rates.get(new Pair(from, to));
 		return rate.intValue();
 	}
